Fix the test for memcg unaware case for uek7
The uek7 list_lru does not have a node entry. The test for memcg aware
is not smart enough to handle the memcg unaware case correctly for uek7.

Add a test for the list_lru helper. It works by walking the dentry list_lru
for one or more fileystems and then looks up the memcg index, and then
verifies the entry can be found by the memcg index. If the filesystem type
in xfs, then walk the xfs_buf list_lru and verify the memcg unaware case.
One can specify verbose output or use the default summary results.

Signed-off-by: Mark Tinguely <mark.tinguely@oracle.com>